Stumb implementation of image.Prepare for darwin.
